voicetype/requirements.txt
pynput
sounddevice
openai
rumps
pyobjc
numpy
scipy
soundfile
py2app
faster-whisper
pyperclip
requests

